rename: page 파일 이름 변경(#128)
