+++ src/App.tsx
@@ -2,14 +2,14 @@ import { BrowserRouter, Route, Routes } from 'react-router';
 import Footer from './features/common/Footer';
 import Header from './features/common/header/Header';
 import HomePage from './pages/HomePage';
-import PlayerPage from './pages/media/FirstPitch';
-import HighlightPage from './pages/media/Highlight';
-import NewsPage from './pages/media/News';
-import PhotoPage from './pages/media/Photo';
-import StoryPage from './pages/media/Story';
+import PlayerPage from './pages/media/FirstPitchPage';
+import HighlightPage from './pages/media/HighlightPage';
+import NewsPage from './pages/media/NewsPage';
+import PhotoPage from './pages/media/PhotoPage';
+import StoryPage from './pages/media/StoryPage';
 
 import WallpaperPage from './pages/ktwiz/WallpaperPage';
-import NewsDetailPage from './pages/media/id/NewsDetail';
+import NewsDetailPage from './pages/media/id/NewsDetailPage';
 import ParkIntroPage from './pages/wizPark/ParkIntroPage';
 import ParkLocationPage from './pages/wizPark/ParkLocationPage';
 import ParkingPage from './pages/wizPark/ParkingPage';
@@ -17,12 +17,12 @@ import ParkingPage from './pages/wizPark/ParkingPage';
 import ScrollToTop from './features/common/ScrollToTop';
 import NotFoundPage from './pages/NotFoundPage';
 import RegularGamePage from './pages/game/regular/RegularGamePage';
-import PressDetailPage from './pages/media/id/PressDetail';
+import PressDetailPage from './pages/media/id/PressDetailPage';
 
 import LoginPage from './pages/auth/LoginPage';
 import SignupPage from './pages/auth/SignupPage';
 import KtHistoryPage from './pages/ktwiz/KtHistoryPage';
-import HighlightDetailPage from './pages/media/id/HighlightDetail';
+import HighlightDetailPage from './pages/media/id/HighlightDetailPage';
 import PlayerListPage from './pages/player/PlayerListPage';
 import TeamMemberDetailPage from './pages/player/detail/TeamMemberDetailPage';
 import IksanStadiumPage from './pages/wizPark/IksanStadiumPage';
@@ -81,6 +81,7 @@ function App() {
             {/* Media 상세 */}
             <Route path="/media/wiznews/:id" element={<NewsDetailPage />} />
             <Route path="/media/wizpress/:id" element={<PressDetailPage />} />
+            <Route path="/media/wizstory/:id" element={<PressDetailPage />} />
             <Route path="/media/photos/:id" element={<PhotoPage />} />
             <Route
               path="/media/highlight/:id"
