+++ src/pages/media/id/NewsDetailPage.tsx
