+++ src/pages/media/id/PressDetailPage.tsx
